+++ src/marketplaces/OpenSeaV1Mapping.ts
@@ -56,7 +56,7 @@ export function handleOSv1Sale(event: OrdersMatched): void {
 
       if (currencyEntity) {
 
-        let amountDecimals = constants.BIGINT_TEN.pow(currencyEntity.decimals)
+        let amountDecimals = constants.BIGINT_TEN.pow(currencyEntity.decimals as u8)
 
         //4. Assign currency address, amount, txId and platform to sale entity
         let saleEntity = new sale(event.block.number.toString() + '-' + event.logIndex.toString())
